v0.1.13 debugging plugin compatibility.
it somehow doesn't see any plugin even if they worked before in the local version. Putting this here with console logs on purpose to see if any of the plugins make it in here or not, is it babel not giving me the plugins or is it this package somehow breaking the plugins? Very interested in getting nesting plugins working with this.