v0.1.17 debug is the only way
this is an effort to see what happens in the component. Need more info as to figure out what's going on, console logs are the way. I'll start from where we get the raw css to when we spit it out to babel (as it's where it seems to break).